requirements.txt
ace_tools==0.0
altgraph==0.17.4
asgiref==3.9.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.2.0
cryptography==45.0.5
cycler==0.12.1
Django==5.2.4
djangorestframework==3.16.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.47.2
fsspec==2025.7.0
gcloud==0.18.3
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numppy==0.0.1
numpy==2.2.6
oauth2client==4.1.3
opencv-python==4.12.0.88
packaging==23.2
pandas==2.2.3
pefile==2023.2.7
pillow==10.2.0
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.1
Pyrebase4==4.8.0
python-dateutil==2.8.2
python-jwt==4.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.4
requests-toolbelt==0.10.1
rsa==4.9.1
scipy==1.14.1
six==1.16.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
Werkzeug==3.1.3
